renamed dhis-support-amqp => artemis, renamed AmqpConfig => ArtemisConfig, configures correct concurrency for topic based CF
