+++ dhis-2/dhis-services/dhis-service-core/pom.xml
@@ -79,7 +79,7 @@
     </dependency>
     <dependency>
       <groupId>org.hisp.dhis</groupId>
-      <artifactId>dhis-support-amqp</artifactId>
+      <artifactId>dhis-support-artemis</artifactId>
     </dependency>
 
     <!-- Other -->
