+++ dhis-2/dhis-services/dhis-service-dxf2/pom.xml
@@ -24,7 +24,7 @@
     </dependency>
     <dependency>
       <groupId>org.hisp.dhis</groupId>
-      <artifactId>dhis-support-amqp</artifactId>
+      <artifactId>dhis-support-artemis</artifactId>
     </dependency>
     <dependency>
       <groupId>org.hisp.dhis</groupId>
