+++ dhis-2/dhis-support/pom.xml
@@ -22,7 +22,7 @@
     <module>dhis-support-system</module>
     <module>dhis-support-jdbc</module>
     <module>dhis-support-expression-parser</module>
-    <module>dhis-support-amqp</module>
+    <module>dhis-support-artemis</module>
   </modules>
 
   <properties>
